refactor: moved login page to seperate component
